Adicionar Mensagem com click e com enter funcionando
